refactor: requets with context (pass headers)
