+++ rest/rest.go
@@ -4,10 +4,8 @@ import (
 	"bytes"
 	"encoding/json"
 	"io"
-	"io/ioutil"
 	"log"
 	"mime/multipart"
-	"net/http"
 	"strconv"
 	"strings"
 
@@ -31,15 +29,6 @@ type Response struct {
 type ResponseResult struct {
 	Context *gin.Context
 	UUID    string
-}
-
-// Request types
-type Request struct {
-	URL     string
-	Method  string
-	Headers map[string]string
-	Body    io.Reader
-	Queries map[string]string
 }
 
 // Log uses current response context to log
@@ -195,54 +184,4 @@ func GetData(jsonBody []byte) (json.RawMessage, error) {
 	}
 	data := body["body"]
 	return data, err
-}
-
-// Send func
-// return []byte, int
-func (request Request) Send() ([]byte, int) {
-	if !ValidMethod(request.Method) {
-		log.Println("[WARN] Unsupported method supplied, use one of constants provided by http package (e.g. http.MethodGet)")
-		return nil, -1
-	}
-
-	req, _ := http.NewRequest(request.Method, request.URL, request.Body)
-
-	for k, v := range request.Headers {
-		req.Header.Set(k, v)
-	}
-
-	if request.Method == http.MethodGet {
-		q := req.URL.Query()
-		for k, v := range request.Queries {
-			q.Add(k, v)
-		}
-		req.URL.RawQuery = q.Encode()
-	}
-
-	client := &http.Client{}
-	resp, err := client.Do(req)
-	if err != nil {
-		log.Println("ERROR: ["+request.Method+"]", err.Error())
-		return nil, -1
-	}
-
-	defer resp.Body.Close()
-	body, _ := ioutil.ReadAll(resp.Body)
-
-	return body, resp.StatusCode
-}
-
-// ValidMethod params
-// @method: string
-// return bool
-func ValidMethod(method string) bool {
-	return method == http.MethodConnect ||
-		method == http.MethodDelete ||
-		method == http.MethodGet ||
-		method == http.MethodHead ||
-		method == http.MethodOptions ||
-		method == http.MethodPatch ||
-		method == http.MethodPost ||
-		method == http.MethodPut ||
-		method == http.MethodTrace
 }
